feat: add mixed precision (fp16) support for training
- Add training.runtime.mixed_precision config key ('float32' default, 'float16' for mixed)
- Apply tf.keras.mixed_precision global policy before model building in main.py
- Force dtype='float32' on output Dense heads (softmax numerical stability)
- Apply mixed precision in HPO worker subprocess runtime (both lock and non-lock paths)
- Add mixed_precision to validation schema
- Enable float16 in production config (e2e_fullscale_production_month1.yaml)

RTX 3090 (Ampere SM 8.6) benefits: ~1.5-2x throughput, ~40-50% less VRAM.
FP8 not feasible (requires Hopper SM 9.0 + TF lacks native support).

+++ models/cnn_lstm_multiclass.py
@@ -192,8 +192,8 @@ def build_cnn_lstm_model(
     num_classes = int(output_cfg["num_classes"])
     output_activation = output_cfg["activation"]
 
-    up_head = layers.Dense(num_classes, activation=output_activation, name="up_intensity")(x)
-    down_head = layers.Dense(num_classes, activation=output_activation, name="down_intensity")(x)
+    up_head = layers.Dense(num_classes, activation=output_activation, name="up_intensity", dtype="float32")(x)
+    down_head = layers.Dense(num_classes, activation=output_activation, name="down_intensity", dtype="float32")(x)
 
     # Build model with appropriate inputs
     if long_term_enabled and long_term_input is not None:
